--- src/PlaceDetails/index.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-const RC: React.FC = () => {
-return (<div>{}</div>);
-};
-export default RC;
