+++ src/lib/ml-form-builder/lib/FieldArray/styles.scss
@@ -24,7 +24,7 @@
 
 .array-add-icon{
     @extend %btn;
-    background-color: #66bf69;
+    // background-color: #66bf69;
     display: flex;
 
 }
